Renamed hookFormReactSpectrum directory to formikReactSpectrum.
