+++ src/view/components/formikReactSpectrum/checkbox.jsx
