+++ src/view/components/formikReactSpectrum/checkbox.styl
